move GDALUseExceptions and gdal_use_exceptions into new utils module
