+++ src/pygeoprocessing/routing/helper_functions.py
@@ -4,7 +4,7 @@ from osgeo import gdal
 from ..geoprocessing import get_raster_info
 from ..geoprocessing import raster_calculator
 from ..geoprocessing_core import DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS
-from ..geoprocessing_core import gdal_use_exceptions
+from ..utils import gdal_use_exceptions
 
 
 @gdal_use_exceptions
