+++ src/pygeoprocessing/routing/watershed.pyx
@@ -26,9 +26,9 @@ import shapely.prepared
 import shapely.wkb
 
 import pygeoprocessing
-from ..geoprocessing_core import gdal_use_exceptions, GDALUseExceptions
 from ..geoprocessing_core import SPARSE_CREATION_OPTIONS
 from ..extensions cimport ManagedRaster
+from ..utils import gdal_use_exceptions, GDALUseExceptions
 
 LOGGER = logging.getLogger(__name__)
